Web-Application/Neural Style Transfer Efficient/requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
smtplib==0.0.1
tensorflow-hub==0.14.0
numpy==1.23.5
Pillow==10.0.0
tensorflow==2.12.0
matplotlib==3.7.2
ipywidgets==8.1.0